Fix some bugs that would attempt to create copies of test directories even if they already existed; add debugging code for parallel image_gen tests

+++ build/copy_cairo.py
@@ -1,5 +1,6 @@
 import platform
 import shutil
+import os
 
 def main():
     python_major, python_minor, _ = platform.python_version_tuple()
@@ -7,7 +8,8 @@ def main():
     long_pyversion = f"python{str(python_major)}.{str(python_minor)}"
     cairo_dir="../cmu_graphics/libs/cairo_loader/modules/cairo_mac/cairo/"
     tox_lib_dir=f"../.tox/{short_pyversion}/lib/{long_pyversion}/site-packages/cairo"
-    shutil.copytree(cairo_dir, tox_lib_dir)
+    if (not os.path.exists(tox_lib_dir)):
+        shutil.copytree(cairo_dir, tox_lib_dir)
 
 if __name__ == "__main__":
     main()
